requirements.txt
django==5.1
psycopg2-binary==2.9.3
djangorestframework
markdown
django-filter
requests